reconcile two different bond referral schemes
git-svn-id: https://svn.salilab.org/imp/trunk@4907 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/atom/src/BondSingletonScore.cpp
@@ -15,7 +15,7 @@ IMPATOM_BEGIN_NAMESPACE
 
 BondSingletonScore::BondSingletonScore(UnaryFunction *f): f_(f){}
 
-Float BondSingletonScore::evaluate(Particle *b,
+double BondSingletonScore::evaluate(Particle *b,
                                    DerivativeAccumulator *da) const
 {
   IMP_IF_CHECK(USAGE_AND_INTERNAL) {
